+++ test/MGNZ.Squidex.CLI.Tests/Routing/AssetNoun.cs
@@ -1,64 +1,25 @@
-namespace MGNZ.Squidex.CLI.Tests.Commands
+namespace MGNZ.Squidex.CLI.Tests.Routing
 {
   using System.Collections.Generic;
 
   using MGNZ.Squidex.CLI.Common.Commands;
   using MGNZ.Squidex.CLI.Common.Routing;
 
-  public class SchemaNoun : Noun
+  public class AssetNoun : Noun
   {
-    public SchemaNoun()
+    public AssetNoun()
     {
-      Names = new[ ] {"schema", "schemas ", "schemata"};
+      Names = new[ ] {"asset", "assets"};
       Verbs = new Dictionary<string, Verb>
       {
-        {
-          "tag", new Verb(this)
-          {
-            Names = new[ ] {"category", "categorise", "tag"},
-            RequestType = typeof(SchemaTagRequest),
-            Options = new Dictionary<string, Option>()
-            {
-              { "name", new Option() {ShortName = "n", LongName = "name", Required = true, OrdanalityOrder = 1} },
-              { "token", new Option() {ShortName = "t", LongName = "tag", Required = true, OrdanalityOrder = 2} },
-              { "alias-credentials", new Option() {ShortName = "c", LongName = "alias-credentials", Required = false} }
-            }
-          }
-        },
         {
           "delete", new Verb(this)
           {
             Names = new[ ] {"delete"},
-            RequestType = typeof(SchemaDeleteRequest),
-            Options = new Dictionary<string, Option>()
-            {
-              { "name", new Option() {ShortName = "n", LongName = "name", Required = true, OrdanalityOrder = 2} },
-              { "alias-credentials", new Option() {ShortName = "c", LongName = "alias-credentials", Required = false} }
-            }
-          }
-        },
-        {
-          "export", new Verb(this)
-          {
-            Names = new[ ] {"export"},
-            RequestType = typeof(SchemaExportRequest),
+            RequestType = typeof(AssetDeleteRequest),
             Options = new Dictionary<string, Option>()
             {
               { "name", new Option() {ShortName = "n", LongName = "name", Required = true, OrdanalityOrder = 1} },
-              { "path", new Option() {ShortName = "p", LongName = "path", Required = true, OrdanalityOrder = 2} },
-              { "alias-credentials", new Option() {ShortName = "c", LongName = "alias-credentials", Required = false} }
-            }
-          }
-        },
-        {
-          "import", new Verb(this)
-          {
-            Names = new[ ] {"import"},
-            RequestType = typeof(SchemaImportRequest),
-            Options = new Dictionary<string, Option>()
-            {
-              { "name", new Option() {ShortName = "n", LongName = "name", Required = true, OrdanalityOrder = 1} },
-              { "path", new Option() {ShortName = "p", LongName = "path", Required = true, OrdanalityOrder = 2} },
               { "alias-credentials", new Option() {ShortName = "c", LongName = "alias-credentials", Required = false} }
             }
           }
@@ -67,9 +28,49 @@ namespace MGNZ.Squidex.CLI.Tests.Commands
           "list", new Verb(this)
           {
             Names = new[ ] {"list"},
-            RequestType = typeof(SchemaListRequest),
+            RequestType = typeof(AssetListRequest),
             Options = new Dictionary<string, Option>()
             {
+              { "alias-credentials", new Option() {ShortName = "c", LongName = "alias-credentials", Required = false} }
+            }
+          }
+        },
+        {
+          "import", new Verb(this)
+          {
+            Names = new[ ] {"import"},
+            RequestType = typeof(AssetImportRequest),
+            Options = new Dictionary<string, Option>()
+            {
+              { "path", new Option() {ShortName = "p", LongName = "path", Required = true, OrdanalityOrder = 1} },
+              { "name", new Option() {ShortName = "n", LongName = "name", Required = true, OrdanalityOrder = 2} },
+              { "token", new Option() {ShortName = "t", LongName = "tags", Required = true, OrdanalityOrder = 3} },
+              { "alias-credentials", new Option() {ShortName = "c", LongName = "alias-credentials", Required = false} }
+            }
+          }
+        },
+        {
+          "export", new Verb(this)
+          {
+            Names = new[ ] {"export"},
+            RequestType = typeof(AssetExportRequest),
+            Options = new Dictionary<string, Option>()
+            {
+              { "name", new Option() {ShortName = "n", LongName = "name", Required = true, OrdanalityOrder = 1} },
+              { "path", new Option() {ShortName = "p", LongName = "path", Required = true, OrdanalityOrder = 2} },
+              { "alias-credentials", new Option() {ShortName = "c", LongName = "alias-credentials", Required = false} }
+            }
+          }
+        },
+        {
+          "tag", new Verb(this)
+          {
+            Names = new[ ] {"tag"},
+            RequestType = typeof(AssetTagRequest),
+            Options = new Dictionary<string, Option>()
+            {
+              { "name", new Option() {ShortName = "n", LongName = "name", Required = true, OrdanalityOrder = 1} },
+              { "token", new Option() {ShortName = "t", LongName = "tags", Required = true, OrdanalityOrder = 2} },
               { "alias-credentials", new Option() {ShortName = "c", LongName = "alias-credentials", Required = false} }
             }
           }
