+++ test/MGNZ.Squidex.CLI.Tests/Routing/SchemaNoun.cs
@@ -1,50 +1,38 @@
-namespace MGNZ.Squidex.CLI.Tests.Commands
+namespace MGNZ.Squidex.CLI.Tests.Routing
 {
   using System.Collections.Generic;
 
   using MGNZ.Squidex.CLI.Common.Commands;
   using MGNZ.Squidex.CLI.Common.Routing;
 
-  public class AssetNoun : Noun
+  public class SchemaNoun : Noun
   {
-    public AssetNoun()
+    public SchemaNoun()
     {
-      Names = new[ ] {"asset", "assets"};
+      Names = new[ ] {"schema", "schemas ", "schemata"};
       Verbs = new Dictionary<string, Verb>
       {
+        {
+          "tag", new Verb(this)
+          {
+            Names = new[ ] {"category", "categorise", "tag"},
+            RequestType = typeof(SchemaTagRequest),
+            Options = new Dictionary<string, Option>()
+            {
+              { "name", new Option() {ShortName = "n", LongName = "name", Required = true, OrdanalityOrder = 1} },
+              { "token", new Option() {ShortName = "t", LongName = "tag", Required = true, OrdanalityOrder = 2} },
+              { "alias-credentials", new Option() {ShortName = "c", LongName = "alias-credentials", Required = false} }
+            }
+          }
+        },
         {
           "delete", new Verb(this)
           {
             Names = new[ ] {"delete"},
-            RequestType = typeof(AssetDeleteRequest),
+            RequestType = typeof(SchemaDeleteRequest),
             Options = new Dictionary<string, Option>()
             {
-              { "name", new Option() {ShortName = "n", LongName = "name", Required = true, OrdanalityOrder = 1} },
-              { "alias-credentials", new Option() {ShortName = "c", LongName = "alias-credentials", Required = false} }
-            }
-          }
-        },
-        {
-          "list", new Verb(this)
-          {
-            Names = new[ ] {"list"},
-            RequestType = typeof(AssetListRequest),
-            Options = new Dictionary<string, Option>()
-            {
-              { "alias-credentials", new Option() {ShortName = "c", LongName = "alias-credentials", Required = false} }
-            }
-          }
-        },
-        {
-          "import", new Verb(this)
-          {
-            Names = new[ ] {"import"},
-            RequestType = typeof(AssetImportRequest),
-            Options = new Dictionary<string, Option>()
-            {
-              { "path", new Option() {ShortName = "p", LongName = "path", Required = true, OrdanalityOrder = 1} },
               { "name", new Option() {ShortName = "n", LongName = "name", Required = true, OrdanalityOrder = 2} },
-              { "token", new Option() {ShortName = "t", LongName = "tags", Required = true, OrdanalityOrder = 3} },
               { "alias-credentials", new Option() {ShortName = "c", LongName = "alias-credentials", Required = false} }
             }
           }
@@ -53,7 +41,7 @@ namespace MGNZ.Squidex.CLI.Tests.Commands
           "export", new Verb(this)
           {
             Names = new[ ] {"export"},
-            RequestType = typeof(AssetExportRequest),
+            RequestType = typeof(SchemaExportRequest),
             Options = new Dictionary<string, Option>()
             {
               { "name", new Option() {ShortName = "n", LongName = "name", Required = true, OrdanalityOrder = 1} },
@@ -63,14 +51,25 @@ namespace MGNZ.Squidex.CLI.Tests.Commands
           }
         },
         {
-          "tag", new Verb(this)
+          "import", new Verb(this)
           {
-            Names = new[ ] {"tag"},
-            RequestType = typeof(AssetTagRequest),
+            Names = new[ ] {"import"},
+            RequestType = typeof(SchemaImportRequest),
             Options = new Dictionary<string, Option>()
             {
               { "name", new Option() {ShortName = "n", LongName = "name", Required = true, OrdanalityOrder = 1} },
-              { "token", new Option() {ShortName = "t", LongName = "tags", Required = true, OrdanalityOrder = 2} },
+              { "path", new Option() {ShortName = "p", LongName = "path", Required = true, OrdanalityOrder = 2} },
+              { "alias-credentials", new Option() {ShortName = "c", LongName = "alias-credentials", Required = false} }
+            }
+          }
+        },
+        {
+          "list", new Verb(this)
+          {
+            Names = new[ ] {"list"},
+            RequestType = typeof(SchemaListRequest),
+            Options = new Dictionary<string, Option>()
+            {
               { "alias-credentials", new Option() {ShortName = "c", LongName = "alias-credentials", Required = false} }
             }
           }
